Extract coverage-specific code out of `compiletest::runtest`
This moves a few hundred lines of coverage-specific code out of the main
module, making navigation a bit easier.
